Refactor `airflow/graph.py` to simplify code-base (#2080)
The goal with this PR is to refactor the `airflow/graph.py` to make the
code more manageable.

The PR #1759 should be merged before the current changes are applied.

The complexity of building DAGs has increased significantly with #1759.
As part of this PR, we ended up duplicating configurations and argument
passing in multiple parts of the code, adding more complexity to the
`airflow/graph.py` module.

This PR also aims to add documentation changes that will help users of
`node_converters` and `node_conversion_by_task_group`.

**Breaking changes**

The following functions are expected to be used internally only to
Cosmos, so we hope these won't impact end-users, but we are documenting
the changes just in case:
- ``generate_task_or_group`` receives ``render_config`` instead of its
individual configurations, such as ``test_behavior``,
``source_rendering_behavior`` and ``enable_owner_inheritance``
- ``create_task_metadata`` receives ``render_config`` instead of its
individual configurations, such as ``test_behavior``,
``source_rendering_behavior`` and ``enable_owner_inheritance``
- ``create_task_metadata`` now expects the ``node_converters`` argument