Fix `ExecutionMode.WATCHER` deadlock in Airflow 3.0 & 3.1 (#2087)
Airflow 3.0 and 3.1 are currently not thread-safe when running
`xcom_push`, which leads to deadlocks.

The problem description and troubleshooting are detailed in ticket
#2057.

Unfortunately, since this issue is not deterministic, I was unable to
create an automated test to reproduce and prevent regressions. That
said, given that the producer node hangs indefinitely until the task
times out, and given that we were no longer able to reproduce the issue
after the current changes, I firmly believe we should proceed with this
change.

Closes: #2057

Co-authored-by: Ash Berlin-Taylor <ash@astronomer.io>