Fix `ExecutionMode.WATCHER` to use `install_dbt_deps` from `ProjectConfig` (#2112)
With the new ExecutionMode.WATCHER, the decision if we install dbt deps
should be done in the same manner as LOCAL or VIRTUALENV execution
modes.

So I am changing the code to use the ProjectConfig settings for dbt deps
to also work with the Watcher and now it will be in sync with the
documentation.

Closes: #2099
(cherry picked from commit ce87c9728f0953a3a29347ecaade67ddf15492f5)