Refactor Watcher `InvocationMode.SUBPROCESS` log parser (#2183)
While the PR #2152 introduced the capability of users running the
`ExecutionMode.WATCHER` with dbt installed in a separate virtualenv, and
still get updates in "real time", it came with two downsides:

1. Spread the watcher logic into multiple areas of the Cosmos code base
(inside the `utils` folder, and also inside the `hooks/subprocess` -
which should be focused only on running suprocess);
2. It changed the subprocess hook to always invoke the method
`_store_dbt_resource_status_from_log(line, **kwargs)` for each log line
made available in stdout, even when users are not using
`ExecutionMode.WATCHER` .

This PR refactors the implementation to keep watcher implementation
details inside the `operators/` folder. It also allows operators that
subclass the `AbstractDbtLocalBase` to define the attribute
`_process_log_line_callable`. Only classes that define this function
will parse the log lines with the desired logic. At the moment, only the
`DbtProducerWatcherOperator` implements the logic of uploading the
status to XCom.