Remove empty variables emission from telemetry metrics (#2197)
Following the change in #2192 that now passes custom metrics using query
params and does not rely on formatting the variables in the URL path
params, we no longer need to keep the variables key that was used to
build the custom metrics. This PR updates telemetry to stop emitting an
empty variables payload and aligns the tests with the new schema.

related: #2192