Remove unnecessary information from telemetry
Information such as dag_id and task_id can have sensitive details, so we are removing them from our telemetry metrics.
This PR also removes the dag_run_id, since it does not give relevant insights for us to decide what to deprecate and what to not deprecate in Cosmos,
one of the  main motivations with the telemetry collection.