Remove unnecessary information from telemetry (#2211)
Information such as `dag_id` and `task_id` can have sensitive details,
so we are removing them from our telemetry metrics.

This PR also removes the dag_run_id, since it does not provide relevant
insights to help us decide what to deprecate and what not to deprecate
in Cosmos, one of the primary motivations for the telemetry collection.