Refactor ExecutionMode.WATCHER so the logic is centralised
In providers, it is common to split Airflow code into operators, triggers, and other folders.
In the case of Cosmos, execution modes were originally added to the operators folder, since all the code for each execution mode was operator-only.
With the introduction of the async and watcher execution modes, the code became more complex, with the need for a custom triggerer for watcher.
With this, watcher execution mode code started spreading throughout the Cosmos code base, making the code distribution more complex than needed.

This PR brings the watcher triggerer logic close to the operator, so that most of the watcher specific logic lives close to the execution mode implementation.

In Cosmos 2.0, we should review how we name Cosmos packages, potentially calling this folder execution_modes.