Break down watcher DAG building logic
Some users mentioned that the producer task would be listed after consumer tasks in the Airflow UI.
We also received the report that sometimes the producer task was scheduled after the consumer tasks, even though it has a higher priority weight.
I discussed these issues with @ashb and he advised we instantiated the producer task before other Airflow tasks while building the task.
This PR aims to accomplish this.