Refactor `ExecutionMode.WATCHER` so the logic is centralised (#2236)
Since https://github.com/astronomer/astronomer-cosmos/pull/2084, the
`ExecutionMode.WATCHER` logic started spreading throughout the Cosmos
code base.

When implementing Airflow providers, it is common practice to split
Airflow code into operators and triggerers folders. However, in the
specific case of Cosmos, I believe this approach makes it harder to read
and maintain the project and that it is better to centralise the
execution mode code.

Historically, Cosmos execution modes were defined in the operators
folder because the code for each execution mode was operator-only. With
the introduction of the async and watcher execution modes, the execution
mode implementation became more complex, requiring a custom triggerer
for the watcher. As a result, the watcher execution mode code began
spreading throughout the Cosmos codebase, complicating code distribution
more than needed.

This PR brings the watcher triggerer logic closer to the operator, so
that most of the watcher-specific logic lives near the execution mode
implementation.

In Cosmos 2.0, we can review how we name the Cosmos execution mode
folder, potentially renaming this folder from operators to
execution_modes.